update: swapped dep to vendor, MAX_BONE_PER_VERTEX can now be modified automatically in shaders

+++ src/renderer/model.cpp
@@ -91,7 +91,7 @@ void Model::draw(ShaderProgram& shader)
     m_mesh.draw(shader);
 }
 
-void Model::update(std::span<glm::mat4> models, double animation_time)
+void Model::update(std::span<glm::mat4> models, float animation_time)
 {
     m_mesh.update_model_ssbos(models);
     if (m_mesh.m_has_bones) {
